+++ Imperium/src/Visualizers/LevelPointVisualizer.cs
@@ -27,13 +27,16 @@ internal class LevelPointVisualizer : BaseVisualizer<bool, Component>
         levelPointParent.transform.SetParent(parent);
     }
 
+    private readonly Color invalidColor = new(1f, 0.47f, 0.07f);
+    private readonly Color validColor = new(0.06f, 1f, 0.64f);
+
     protected override void OnRefresh(bool isSceneLoaded)
     {
         ClearObjects();
 
         if (!isSceneLoaded) return;
 
-        var connections = new HashSet<ValueTuple<Vector3, Vector3>>();
+        var connections = new Dictionary<Edge, LineRenderer>();
         foreach (var levelPoint in Object.FindObjectsByType<LevelPoint>(FindObjectsSortMode.None))
         {
             var levelPointIndicator = Object.Instantiate(ImpAssets.LevelPoint, levelPointParent.transform);
@@ -44,22 +47,25 @@ internal class LevelPointVisualizer : BaseVisualizer<bool, Component>
 
             foreach (var connectedPoint in levelPoint.ConnectedPoints)
             {
-                var connection = new ValueTuple<Vector3, Vector3>(
-                    levelPoint.transform.position, connectedPoint.transform.position
-                );
-                if (connections.Contains(connection)) continue;
+                var connection = new Edge(levelPoint.transform.position, connectedPoint.transform.position);
+                if (!connections.TryGetValue(connection, out var connectionLine))
+                {
+                    connectionLine = Geometry.CreateLine(
+                        levelPointParent.transform,
+                        useWorldSpace: true,
+                        startColor: invalidColor,
+                        endColor: invalidColor,
+                        thickness: 0.025f,
+                        positions: [connection.PointA + Vector3.up * 0.04f, connection.PointB + Vector3.up * 0.04f]
+                    );
 
-                var connectionLine = Geometry.CreateLine(
-                    levelPointParent.transform,
-                    useWorldSpace: true,
-                    startColor: new Color(0.06f, 1f, 0.64f),
-                    endColor: new Color(0.06f, 1f, 0.64f),
-                    thickness: 0.025f,
-                    positions: [connection.Item1 + Vector3.up * 0.04f, connection.Item2 + Vector3.up * 0.04f]
-                );
-
-                visualizerObjects.Add(connectionLine.GetInstanceID(), connectionLine.transform);
-                connections.Add(connection);
+                    visualizerObjects.Add(connectionLine.GetInstanceID(), connectionLine.transform);
+                    connections[connection] = connectionLine;
+                }
+                else
+                {
+                    Geometry.SetLineColor(connectionLine, validColor, validColor);
+                }
             }
         }
     }